JWT Auth implemented

+++ Project/ESD-Backend/Student-Fee-Service/src/main/java/com/enum3rat3/studentfeeservice/controller/BillsController.java
@@ -3,14 +3,14 @@ package com.enum3rat3.studentfeeservice.controller;
 import com.enum3rat3.studentfeeservice.model.*;
 import com.enum3rat3.studentfeeservice.service.BillsService;
 import com.enum3rat3.studentfeeservice.service.StudentBillsService;
-import jakarta.persistence.criteria.CriteriaBuilder;
+import com.enum3rat3.studentfeeservice.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("api/bills")
+@RequestMapping("/api/bills")
 public class BillsController {
     @Autowired
     private BillsService billsService;
@@ -18,21 +18,52 @@ public class BillsController {
     @Autowired
     private StudentBillsService studentBillsService;
 
+    @Autowired
+    private UserService userService;
+
     @PostMapping("create")
-    private String createBill(@RequestBody StudentBillDetails studentBillsDetails) {
+    private String createBill(@RequestBody StudentBillDetails studentBillsDetails, @RequestHeader("Authorization") String jwt) throws Exception{
+
+        if(jwt == null){
+            throw new Exception("jwt required...");
+        }
+        User user = userService.getUserProfileHandler(jwt);
+
+        if(user == null){
+            throw new Exception("UNAUTHORIZED CREDENTIALS");
+        }
+
         billsService.createBill(studentBillsDetails);
         return "Bill created";
     }
 
     @GetMapping("read/{studentId}")
-    private List<Bills> readBills(@PathVariable String studentId) {
-        List<Integer> billIDs = studentBillsService.readBills(studentId);
+    private List<Bills> readBills(@PathVariable String studentId, @RequestHeader("Authorization") String jwt) throws Exception{
+        if(jwt == null){
+            throw new Exception("jwt required...");
+        }
+        User user = userService.getUserProfileHandler(jwt);
 
+        if(user == null){
+            throw new Exception("UNAUTHORIZED CREDENTIALS");
+        }
+
+        List<Integer> billIDs = studentBillsService.readBills(studentId);
         return billsService.readBills(billIDs);
     }
 
     @PutMapping("update/{studentId}/{billId}")
-    private String updateBill(@PathVariable String studentId, @PathVariable int billId, @RequestBody StudentBillDetails studentBillDetails) {
+    private String updateBill(@PathVariable String studentId, @PathVariable int billId, @RequestBody StudentBillDetails studentBillDetails, @RequestHeader("Authorization") String jwt) throws Exception{
+
+        if(jwt == null){
+            throw new Exception("jwt required...");
+        }
+        User user = userService.getUserProfileHandler(jwt);
+
+        if(user == null){
+            throw new Exception("UNAUTHORIZED CREDENTIALS");
+        }
+
         if(billsService.updateBill(studentId, billId, studentBillDetails))
         {
             return "Bill updated for " + studentId;
@@ -41,7 +72,16 @@ public class BillsController {
     }
 
     @DeleteMapping("delete/{studentId}/{billId}")
-    private String deleteBill(@PathVariable String studentId, @PathVariable int billId) {
+    private String deleteBill(@PathVariable String studentId, @PathVariable int billId, @RequestHeader("Authorization") String jwt) throws Exception {
+
+        if(jwt == null){
+            throw new Exception("jwt required...");
+        }
+        User user = userService.getUserProfileHandler(jwt);
+
+        if(user == null){
+            throw new Exception("UNAUTHORIZED CREDENTIALS");
+        }
         if(billsService.deleteBill(studentId, billId))
         {
             return "Bill details deleted for " + studentId;
@@ -51,15 +91,31 @@ public class BillsController {
 
     // ********************* Domain Related Operation ******************
     @PostMapping("create/domain/")
-    private String createBill(@RequestBody DomainBillDetails domainBillDetails) {
-        billsService.createBillForDomain(domainBillDetails);
+    private String createBill(@RequestBody DomainBillDetails domainBillDetails, @RequestHeader("Authorization") String jwt) throws Exception{
 
+        if(jwt == null){
+            throw new Exception("jwt required...");
+        }
+        User user = userService.getUserProfileHandler(jwt);
+
+        if(user == null){
+            throw new Exception("UNAUTHORIZED CREDENTIALS");
+        }
+        billsService.createBillForDomain(domainBillDetails);
         return "Bill created for " + domainBillDetails.getDomainId();
     }
 
     @GetMapping("read/domain/{domainId}")
-    private List<Bills> readBillsForDomain(@PathVariable int domainId) {
+    private List<Bills> readBillsForDomain(@PathVariable int domainId, @RequestHeader("Authorization") String jwt) throws Exception{
+
+        if(jwt == null){
+            throw new Exception("jwt required...");
+        }
+        User user = userService.getUserProfileHandler(jwt);
+
+        if(user == null){
+            throw new Exception("UNAUTHORIZED CREDENTIALS");
+        }
         return billsService.readBillsForDomain(domainId);
     }
-
 }
